requirements.txt
# ID: REQ-001-V01
# IdeaScrawl Agenda - Backend Dependencies

# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Storage
supabase==2.17.0
psycopg2-binary==2.9.9

# Image Processing (for local fallback)
Pillow==10.1.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
uuid==1.30

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS for web dashboard  
# python-cors==1.7.0  # Not needed, FastAPI has built-in CORS